https://semantic-kitti.org/dataset.html
Overview
We present a large-scale dataset based on the KITTI Vision Benchmark and we used all sequences provided by the odometry task. We provide dense annotations for each individual scan of sequences 00-10, which enables the usage of multiple sequential scans for semantic scene interpretation, like semantic segmentation and semantic scene completion.

The remaining sequences, i.e., sequences 11-21, are used as a test set showing a large variety of challenging traffic situations and environment types. Labels for the test set are not provided and we use an evaluation service that scores submissions and provides test set results.